requirements.txt
appdirs==1.4.4
attrs==19.3.0
backcall==0.2.0
black==19.10b0
bleach==3.1.5
cfgv==3.1.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
depq==1.5.5
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
identify==1.4.25
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupytext==1.5.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nodeenv==1.4.0
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
pre-commit==2.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.7.14
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
virtualenv==20.0.27
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
